Connection should be in try catch block...
Currently If redis server down or something error. Current page generation will fail and return 500.
we should put connection function inside try catch block. So It will still continue without the page cache. and record in in error_log

To reproduce. Just change the ip or port of the redis.